@@ -2,13 +2,15 @@ apply plugin: 'com.android.library'
 
 
 android {
-    compileSdkVersion 26
+    compileSdk 30
+    buildToolsVersion '30.0.3'
     defaultConfig {
         minSdkVersion 19
         //noinspection ExpiredTargetSdkVersion
-        targetSdkVersion 22
+        targetSdkVersion 30
 
-        buildConfigField 'String', 'VA_VERSION_NAME', '"V' + rootProject.ext.VERSION + '"'
+        buildConfigField 'String', 'VA_VERSION_NAME', '"V' + rootProject.ext.VA_VERSION + '"'
+        buildConfigField 'int', 'VERSION_CODE', '' + rootProject.ext.VERSION_CODE + ''
 
         if (project.hasProperty('commit')) {
             buildConfigField 'String', 'commit', '"' + commit + '"'
@@ -45,6 +47,7 @@ android {
         abortOnError false
         warning 'NewApi', 'OnClick', 'EllipsizeMaxLines'
     }
+    ndkVersion '28.0.12433566 rc1'
 }
 gradle.taskGraph.afterTask { task ->
     def TASK_NAME = "transformNativeLibsWithStripDebugSymbolFor"
@@ -77,6 +80,7 @@ gradle.taskGraph.afterTask { task ->
                         toFile.getParentFile().mkdirs()
                         toFile.createNewFile()
                     }
+//                    soFile.delete()
                     println("copy : " + soFile.path)
                 }
             }
@@ -85,17 +89,20 @@ gradle.taskGraph.afterTask { task ->
                 if (soFile.name.endsWith("v++.so")) {
                     def toFile = new File(numberDir, "lib/arm64-v8a/" + soFile.name)
                     GFileUtils.copyFile(soFile, toFile)
+//                    soFile.delete()
                     println("copy : " + soFile.path)
                 }
             }
+//            soDir.delete()
         }
     }
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    compileOnly 'com.xdja.safekeyservice:xdjacrypto:1.0.3'
-    compileOnly 'com.xdja.safekeyservice:jar_multi_jniapi:3.9.50'
-    compileOnly 'com.xdja.safekeyservice:jar_multi_unitepin:3.9.24'
+    compileOnly fileTree(include: ['*.aar'], dir: 'libs')
+//    compileOnly 'com.xdja.safekeyservice:xdjacrypto:1.0.3'
+//    compileOnly 'com.xdja.safekeyservice:jar_multi_jniapi:3.9.50'
+//    compileOnly 'com.xdja.safekeyservice:jar_multi_unitepin:3.9.24'
     implementation 'com.android.support:appcompat-v7:26.1.0'
 }
