@@ -48,7 +48,7 @@ import mirror.android.net.wifi.WifiSsid;
 
 /**
  * @author Lody
- * @see android.net.wifi.WifiManager
+ * @see WifiManager
  */
 public class WifiManagerStub extends BinderInvocationProxy {
 
