@@ -11,7 +11,7 @@ public class NotificationL {
     public static Class<?> TYPE = RefClass.load(NotificationL.class, Notification.class);
 
     public static class Builder {
-        public static Class<?> TYPE = RefClass.load(Builder.class, android.app.Notification.Builder.class);
+        public static Class<?> TYPE = RefClass.load(Builder.class, Notification.Builder.class);
 
         @MethodParams({Context.class, Notification.class})
         public static RefStaticMethod<Notification> rebuild;
