@@ -11,7 +11,7 @@ public class INfcAdapter {
     public static Class<?> TYPE = RefClass.load(INfcAdapter.class, "android.nfc.INfcAdapter");
 
     public static class Stub {
-        public static Class<?> TYPE = RefClass.load(INfcAdapter.Stub.class, "android.nfc.INfcAdapter$Stub");
+        public static Class<?> TYPE = RefClass.load(Stub.class, "android.nfc.INfcAdapter$Stub");
         @MethodParams({IBinder.class})
         public static RefStaticMethod<IInterface> asInterface;
     }
