@@ -14,7 +14,7 @@ public class IAccessibilityManager {
     public static Class<?> TYPE = RefClass.load(IAccessibilityManager.class, "android.view.accessibility.IAccessibilityManager");
 
     public static class Stub {
-        public static Class<?> TYPE = RefClass.load(IAccessibilityManager.Stub.class, "android.view.accessibility.IAccessibilityManager$Stub");
+        public static Class<?> TYPE = RefClass.load(Stub.class, "android.view.accessibility.IAccessibilityManager$Stub");
         @MethodParams({IBinder.class})
         public static RefStaticMethod<IInterface> asInterface;
     }
